melhoria upgrades

+++ Projeto 5/Assets/Scripts/StateNameController.cs
@@ -11,7 +11,7 @@ public class StateNameController : MonoBehaviour
     public static int snccoinqt;
     public static int snchp;
     public static string snc_scene;
-    public static int sncfirerate;
+    public static float sncfirerate;
     public static int sncdamage;
     public static int sncfirepoint;
 }
